🔨 refactor: 팝업 포커스 개선
